Allow writing multi-valued parameters
This changeset allows to use the nova parameter type “MultiStrOpt”
which allow to declare a parameter multiple times with different values.

It's a bit broken as reading a config file with a parameter declared
multiple times will raise a warning :
“warning: Nova_config notification_driver found in both configfile and
configfile; skipping the configfile version”

Also, using this feature without purging the nova_config resources will
lead to duplicates of the first value found for the parameter(s) defined
multiple times.